+++ lib/flowbite/view_components.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "view_components/engine"
 require_relative "view_components/version"
 
 module Flowbite