+++ lib/flowbite/view_components/engine.rb
@@ -1,0 +1,13 @@
+require "rails/engine"
+
+module Flowbite
+  module ViewComponents
+    class Engine < ::Rails::Engine
+      isolate_namespace Flowbite::ViewComponents
+
+      config.autoload_paths = [
+        "#{root}/app/components"
+      ]
+    end
+  end
+end